+++ discovery-provider/scripts/dev-server.sh
@@ -4,7 +4,7 @@
 # Exports environment variables necessary for Flask app
 
 # app entry point module
-export FLASK_APP=src
+export FLASK_APP=src.app
 # enables flask development environment, including interactive debugger and reloader
 # - (http://flask.pocoo.org/docs/1.0/server/)
 export FLASK_ENV=development
