+++ discovery-provider/src/tasks/index_network_peers.py
@@ -1,7 +1,7 @@
 import logging
 import concurrent.futures
 from src.tasks.celery_app import celery
-from src import eth_abi_values
+from src.app import eth_abi_values
 from src.utils.helpers import get_ipfs_info_from_cnode_endpoint, is_fqdn
 from src.models import User
 from src.utils.redis_cache import pickle_and_set, get_sp_id_key, get_pickled_key
